change blink input to left control and format auto format
