+++ Assets/Prototyping/ChargeJumpPrototype.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 26585db4124de674981ded1a02bd1e6c
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
